requirements.txt
pydantic>=2.0.0
pytest>=7.0.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0
